requirements.txt
altgraph==0.17.4
asgiref==3.10.0
bsdiff4==1.2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
macholib==1.16.3
packaging==25.0
pillow==12.0.0
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyNaCl==1.6.0
requests==2.32.5
securesystemslib==0.31.0
setuptools==80.9.0
sqlparse==0.5.3
tuf==4.0.0
tufup==0.10.0
urllib3==2.5.0
waitress==3.0.2
whitenoise==6.11.0
